+++ CHANGELOG.rst
@@ -23,7 +23,10 @@ Changed
 ^^^^^^^
 
 - updated WeightsColumnMixin to use new narwhals 'is_finite' method
-- placeholder
+- narwhalified ModeImputer `#321 <https://github.com/lvgig/tubular/issues/321>_`
+- fixed issues with all null and nullable-bool column handling in dataframe_init_dispatch
+- added NaN error handling to WeightColumnMixin
+- narwhalified MeanImputer `#344 https://github.com/lvgig/tubular/issues/344_`
 - placeholder
 - placeholder
 - placeholder